src/requirements.txt
adversarial-robustness-toolbox
numpy 
pandas
scikit-learn
tensorflow
keras
flask
zope.interface
lightgbm
xgboost
a2pm
flask-cors
catboost